pyproject.toml
[tool.poetry]
name = "core-africare-identity"
version = "0.1.0"
description = "Identity management and Keycloak integration"
authors = ["Africare Team <team@africare.app>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
# PostgreSQL dependencies
alembic = "^1.16.5"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.0"
fastapi = "^0.118.0"
fastapi-errors-rfc9457 = {git = "https://github.com/btall/fastapi-errors-rfc9457.git", tag = "v0.1.0", extras = ["opentelemetry"]}
pydantic = {version = ">=2.9.0,<2.11.0", extras = ["email"]}
pydantic-settings = "^2.10.0"
python-dotenv = "^1.1.1"
python-keycloak = "^5.8.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
uvicorn = "^0.37.0"
opentelemetry-api = "^1.37"
opentelemetry-sdk = "^1.37"
opentelemetry-distro = "^0.58b0"
opentelemetry-exporter-otlp = "^1.37"
opentelemetry-exporter-otlp-proto-grpc = "^1.37"
opentelemetry-instrumentation-fastapi = "^0.58b0"
opentelemetry-instrumentation-httpx = "^0.58b0"
opentelemetry-instrumentation = "^0.58b0"
tenacity = "^9.1.2"
# Redis messaging (Phase 1 MVP)
redis = {extras = ["hiredis"], version = "^5.2.0"}
bcrypt = "^5.0.0"
# FHIR integration (HAPI FHIR) - v8+ uses Pydantic v2 natively
fhir-resources = "^8.1.0"
httpx = "^0.28.0"
[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.2.0"
pytest-timeout = "^2.3.1"
ruff = "^0.14.0"
mypy = "^1.18.2"
alembic = "^1.16.5"
gevent = "^25.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 100
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B",
    "UP",
    "N",
    "RUF",
]
ignore = [
    "E501",
    "B008",
    "UP006",
    "E402",
    "B904",
]
fixable = ["ALL"]
unfixable = []
exclude = [
    "*.pyc",
    ".git",
    ".ruff_cache",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]
# keycloak_sync_service a des handlers webhook complexes par nature
"app/services/keycloak_sync_service.py" = ["C901"]
